+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Changed
   - Upgrade minimum Python to 3.10. (GitLab #62)
+  - Replaced switch with new match syntax. (GitLab #62)
 
 <!-- This comment will prevent merge conflicts -->
 ## Release History